# docs/openapi-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LogicPOS.ApiServer | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001/"
    }
  ],
  "paths": {
...
    "/articles/{id}": {
      "get": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/articles/{id}": {
      "get": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/articles/stocks/warehouses/{id}": {
      "put": {
        "tags": [
          "Stocks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarehouseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarehouseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarehouseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Stocks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/articles/stocks/warehouses/{id}": {
      "put": {
        "tags": [
          "Stocks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarehouseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarehouseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarehouseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Stocks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/articles/warehouse-articles": {
      "get": {
        "tags": [
          "WarehouseArticles"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "articleId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/articles/warehouse-articles": {
      "get": {
        "tags": [
          "WarehouseArticles"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "articleId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/articles/warehouse-articles/{id}": {
      "get": {
        "tags": [
          "WarehouseArticles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "WarehouseArticles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/articles/warehouse-articles/{id}": {
      "get": {
        "tags": [
          "WarehouseArticles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "WarehouseArticles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ArticlePriceRequest": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "format": "double"
          },
          "promotionValue": {
            "type": "number",
            "format": "double"
          },
          "usePromotion": {
            "type": "boolean"
          }
        },
        "nullable": true
      },
      "ButtonResponse": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "imageExtension": {
            "type": "string"
          }
        },
        "nullable": true
      },
...
      "UpdateArticleRequest": {
        "type": "object",
        "properties": {
          "order": {
            "type": "integer",
            "format": "uint32"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "codeDealer": {
            "type": "string",
            "nullable": true
          },
          "designation": {
            "type": "string"
          },
          "button": {
            "$ref": "#/components/schemas/ButtonResponse"
          },
          "price1": {
            "$ref": "#/components/schemas/ArticlePriceRequest"
          },
          "price2": {
            "$ref": "#/components/schemas/ArticlePriceRequest"
          },
          "price3": {
            "$ref": "#/components/schemas/ArticlePriceRequest"
          },
          "price4": {
            "$ref": "#/components/schemas/ArticlePriceRequest"
          },
          "price5": {
            "$ref": "#/components/schemas/ArticlePriceRequest"
          },
          "priceWithVat": {
            "type": "boolean"
          },
          "discount": {
            "type": "number",
            "format": "double"
          },
          "defaultQuantity": {
            "type": "number",
            "format": "double"
          },
          "totalStock": {
            "type": "number",
            "format": "double"
          },
          "minimumStock": {
            "type": "number",
            "format": "double"
          },
          "tare": {
            "type": "number",
            "format": "double"
          },
          "weight": {
            "type": "number",
            "format": "float"
          },
          "barcode": {
            "type": "string",
            "nullable": true
          },
          "pvpVariable": {
            "type": "boolean"
          },
          "favorite": {
            "type": "boolean"
          },
          "useWeighingBalance": {
            "type": "boolean"
          },
          "subfamilyId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "typeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "classId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "measurementUnitId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "sizeUnitId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "commissionGroupId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "discountGroupId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "vatOnTableId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "vatDirectSellingId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "vatExemptionReasonId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "isComposed": {
            "type": "boolean"
          },
          "uniqueArticles": {
            "type": "boolean"
          },
          "isSdrPackaging": {
            "type": "boolean"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "barcodeLabelPrintModel": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "UpdateWarehouseRequest": {
        "type": "object",
        "properties": {
          "order": {
            "type": "integer",
            "format": "uint32"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "designation": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  },
...
}